requirements.txt
# Feel free to add your Python dependencies below. For formatting guidelines, see:
# https://pip.pypa.io/en/latest/reference/requirements-file-format/

apify < 3.0
pydantic_ai >= 0.0.29
PyGithub >= 2.6.1
